Changed acceleration type to float in activityClassifier.h.

+++ src/boot/ksdk1.1.0/activityClassifier.h
@@ -39,11 +39,11 @@ int16_t ZAcceleration = 0;
 detectedActivity activityReading;
 
 // Define buffers to store the magnitudes of the acceleration measurements before and after low-pass filtering.
-uint16_t AccelerationBuffer[BUFFER_SIZE] = {0}; // Initialised to 0.
-uint32_t LPFBuffer[BUFFER_SIZE] = {0}; // Initialised to 0.
-uint16_t absXAcceleration, absYAcceleration, absZAcceleration; // Absolute values of the X, Y and Z acceleration measurements.
-uint16_t maximalAcceleration; // Maximum value among the X, Y and Z acceleration measurements.
-uint32_t maximumValue, minimumValue, LPFBufferMidpoint, numberOfCrossings; // Not set until the classification algorithm runs.
+float AccelerationBuffer[BUFFER_SIZE] = {0}; // Initialised to 0.
+float LPFBuffer[BUFFER_SIZE] = {0}; // Initialised to 0.
+float absXAcceleration, absYAcceleration, absZAcceleration; // Absolute values of the X, Y and Z acceleration measurements.
+float maximalAcceleration; // Maximum value among the X, Y and Z acceleration measurements.
+float maximumValue, minimumValue, LPFBufferMidpoint, numberOfCrossings; // Not set until the classification algorithm runs.
 
 uint16_t numberOfSteps = 0; // Cumulative number of steps since booting the device. Initialised to 0.
 float speed = 0; // Estimated speed in km/hr. Initialised to 0.
